[refactor] Move methods to separate file
